Move run-in-container jobs to run on docker-cloud
The docker-custom-build-environment-plugin doesn't work in a
docker-in-docker agent, so it doesn't work either running agents on a
docker-master nor a docker-agent, because it bind-mounts the agent
folders into the sub-container, and because bind-mounts are
host->container and not container->container, the sub-container doesn't
get the expected files and doesn't work.

This works around this by launching the right container directly, and
letting jenkins agent drop the relevant files into that container.

Some jobs blindly start other docker containers, and binds ports to
random local ports. Those random local ports are in the host context, so
run.sh rewrites /etc/hosts to point to the gateway machine and that way
it looks to the containers as if the random local port appeared on
localhost.