Allow in-jenkins overwrite of how dsl acts
This allows simple overwrite and filtering of what to generate jobs for
and how to act.

This way, one does not need to hack github_docker_repos.groovy to test
changes to .jenkins.yaml files, just copy the existing
bootstrap-docker-builds job and give it some extra environment variables
to change its behavior.