Migrate from docker-build-publish to docker-plugin
The docker-build-publish plugin isn't supported to be run from pipeline,
but the docker-plugin is.

Also, the docker-build-publish plugin is pretty dead, and haven't had
any activity for the last 3 years, so migrating away from it to the
actiely developed docker-plugin feels like the right thing to do.

Some minor functionallity from docker-build-publish isn't implemented in
the docker-plugin, but it feels like we can live without it for now.